Use float instead of double (no need in such precision)

+++ FuzzyString/JaccardDistance.cs
@@ -5,14 +5,14 @@ namespace FuzzyString
 {
     public static partial class ComparisonMetrics
     {
-        public static double JaccardDistance(this string source, string target)
+        public static float JaccardDistance(this string source, string target)
         {
             return 1 - source.JaccardIndex(target);
         }
 
-        public static double JaccardIndex(this string source, string target)
+        public static float JaccardIndex(this string source, string target)
         {
-            return Convert.ToDouble(source.Intersect(target).Count()) / Convert.ToDouble(source.Union(target).Count());
+            return Convert.ToSingle(source.Intersect(target).Count()) / Convert.ToSingle(source.Union(target).Count());
         }
     }
 }
